Rename IUserPreferencesContext.tsx -> UserPreferencesProvider.tsx
Move UserPreferencesProvider.tsx
 to providers
